Separate payload patch into another file
